Format code files

+++ Issues/3_Exception_Error_Handling_Disorders/3B_Unexpected_Throw_Or_Revert/Unexpected_Throw_Or_Revert.sol
@@ -3,21 +3,16 @@ pragma solidity 0.8.17;
 
 //Once a higher bid is submitted, the previous bid is returned to the corresponding bidder.
 //However, if the previous bidder is a contract with a complex callback, the "send" operation will never succed. Thus, his bid will never be replaced.
-contract UnexpectedThrowOrRevert{
-
+contract UnexpectedThrowOrRevert {
     address private _highestBidder;
-    uint private _highestBid;
+    uint256 private _highestBid;
 
-
-    function Bid() external payable{
-        if(msg.value>_highestBid)
-        {
+    function Bid() external payable {
+        if (msg.value > _highestBid) {
             require(payable(_highestBidder).send(_highestBid));
-            _highestBidder=msg.sender;
-            _highestBid=msg.value;
-        }
-        else
-        {
+            _highestBidder = msg.sender;
+            _highestBid = msg.value;
+        } else {
             revert("The big was lower than the current highest bid");
         }
     }
