+++ Issues/3_Exception_Error_Handling_Disorders/3B_Unexpected_Throw_Or_Revert/Unexpected_Throw_Or_RevertFixed.sol
@@ -2,30 +2,27 @@
 pragma solidity 0.8.17;
 
 //Fix: Implement a "pull" rather than "push" payment via a cashback method.
-contract UnexpectedThrowOrRevertFixed{
-
+contract UnexpectedThrowOrRevertFixed {
     address private _highestBidder;
-    uint private _highestBid;
+    uint256 private _highestBid;
 
-    mapping (address=>uint) owedFormerHighestBidders;
+    mapping(address => uint256) owedFormerHighestBidders;
 
-    function Bid() external payable
-    {
-        if(msg.value>_highestBid)
-        {
-            owedFormerHighestBidders[_highestBidder]+=_highestBid;
-            _highestBidder=msg.sender;
-            _highestBid=msg.value;
-        }
-        else
-        {
+    function Bid() external payable {
+        if (msg.value > _highestBid) {
+            owedFormerHighestBidders[_highestBidder] += _highestBid;
+            _highestBidder = msg.sender;
+            _highestBid = msg.value;
+        } else {
             revert("The big was lower than the current highest bid");
         }
     }
 
-    function Cashback() external
-    {
-        require(owedFormerHighestBidders[msg.sender]>0,"You have nothing to collect");
+    function Cashback() external {
+        require(
+            owedFormerHighestBidders[msg.sender] > 0,
+            "You have nothing to collect"
+        );
         payable(msg.sender).transfer(owedFormerHighestBidders[msg.sender]);
     }
 }
