+++ Issues/3_Exception_Error_Handling_Disorders/3D_Assert_Require_Revert_Violation/AssertAlwaysFalse.sol
@@ -1,11 +1,13 @@
 // SPDX-License-Identifier: MIT
 pragma solidity 0.8.17;
 
-contract AssertAlwaysFalse{
-    
-    function notGonnaExecute(uint parameter) external pure returns(uint)
+contract AssertAlwaysFalse {
+    function notGonnaExecute(uint256 parameter)
+        external
+        pure
+        returns (uint256)
     {
-        require(parameter<0); // this require always fails because paramenter is uint
+        require(parameter < 0); // this require always fails because paramenter is uint
         return parameter;
     }
 }
