+++ Issues/3_Exception_Error_Handling_Disorders/3D_Assert_Require_Revert_Violation/AssertAlwaysFalseDefaultValueFixed.sol
@@ -2,21 +2,18 @@
 pragma solidity 0.8.17;
 
 //Fix: The propser comparison operator is used
-contract AssertAlwaysFalse{
-    
-    uint private _limit;
-    uint private _currentCount;
-    mapping(uint=>address) private _participants;
+contract AssertAlwaysFalse {
+    uint256 private _limit;
+    uint256 private _currentCount;
+    mapping(uint256 => address) private _participants;
 
-    constructor(uint limit)
-    {
-        _limit=limit;
+    constructor(uint256 limit) {
+        _limit = limit;
     }
 
-    function register() external
-    {
-        require(_limit>_currentCount,"The limit has been reached");
-        _currentCount+=1;
-        _participants[_currentCount]=msg.sender;      
+    function register() external {
+        require(_limit > _currentCount, "The limit has been reached");
+        _currentCount += 1;
+        _participants[_currentCount] = msg.sender;
     }
 }
