+++ Issues/3_Exception_Error_Handling_Disorders/3D_Assert_Require_Revert_Violation/AssertAlwaysFalseEqualityComparison.sol
@@ -2,12 +2,14 @@
 pragma solidity 0.8.17;
 
 //The method will never return since there is a contradiction in the require statement, a variable cannot be different from the value that was just assigned to it.
-contract AssertAlwaysFalse{
-    
-    function notGonnaExecute(uint parameter1,uint parameter2) external pure returns(uint)
+contract AssertAlwaysFalse {
+    function notGonnaExecute(uint256 parameter1, uint256 parameter2)
+        external
+        pure
+        returns (uint256)
     {
-        uint number=parameter1*parameter2;
-        require(number!=parameter1*parameter2);
+        uint256 number = parameter1 * parameter2;
+        require(number != parameter1 * parameter2);
         return number;
     }
 }
