+++ Issues/3_Exception_Error_Handling_Disorders/3D_Assert_Require_Revert_Violation/AssertAlwaysTrueUintPositive.sol
@@ -2,11 +2,13 @@
 pragma solidity 0.8.17;
 
 //This contract contains a tautology in its "require", an unsigned integer will always have a value higher than 0
-contract AssertAlwaysTrue{
-    
-    function uselessAssertUint(uint parameter) external pure returns(uint)
+contract AssertAlwaysTrue {
+    function uselessAssertUint(uint256 parameter)
+        external
+        pure
+        returns (uint256)
     {
-        require(parameter>=0);
+        require(parameter >= 0);
         return parameter;
     }
 }
