+++ Issues/4_Denial_of_Service/4A_LockedEther/LockedEtherFixed.sol
@@ -3,18 +3,18 @@ pragma solidity 0.8.17;
 
 //Fix: Define a function to transfer the Ether
 contract LockedEtherFixed {
-    mapping(address => uint) private _balance;
+    mapping(address => uint256) private _balance;
 
-    function deposit(uint amount) external payable {
+    function deposit(uint256 amount) external payable {
         require(msg.value == amount, "Incorrect amount");
         _balance[msg.sender] = amount;
     }
 
     //Obs: slither reentracy false positive
-     function withdraw() external {
-        uint toSend=_balance[msg.sender];
-        require(_balance[msg.sender]!=0,"No balance found");
-        _balance[msg.sender]=0;
+    function withdraw() external {
+        uint256 toSend = _balance[msg.sender];
+        require(_balance[msg.sender] != 0, "No balance found");
+        _balance[msg.sender] = 0;
         bool sent = payable(msg.sender).send(toSend);
         require(sent, "Failed to send Ether");
     }
