+++ Issues/7_Arithmetic_Bugs/7B_Divide_By_Zero/DivideByZeroDirect.sol
@@ -3,8 +3,11 @@ pragma solidity 0.8.17;
 
 //This contract performs a division where the divisor is always zero.
 contract DivideByZeroMinimal {
-
-    function divide(uint number1,uint number2) public pure returns(uint)  {
-        return number1/(number2-number2);
+    function divide(uint256 number1, uint256 number2)
+        public
+        pure
+        returns (uint256)
+    {
+        return number1 / (number2 - number2);
     }
 }
