+++ Issues/7_Arithmetic_Bugs/7B_Divide_By_Zero/DivideByZeroMinimal.sol
@@ -3,8 +3,11 @@ pragma solidity 0.8.17;
 
 //This contract contains a divison by an outside parameter that is not zero checked.
 contract DivideByZeroMinimal {
-
-    function divide(uint number,uint divisor) external pure returns(uint) {
-        return number/divisor;
+    function divide(uint256 number, uint256 divisor)
+        external
+        pure
+        returns (uint256)
+    {
+        return number / divisor;
     }
 }
