+++ Issues/7_Arithmetic_Bugs/7B_Divide_By_Zero/DivideByZeroMinimalFixed.sol
@@ -3,9 +3,12 @@ pragma solidity 0.8.17;
 
 //Fix: Check that the divisor is not zero
 contract DivideByZeroMinimalFixed {
-
-    function divide(uint number,uint divisor) external pure returns(uint) {
-        require(divisor>0,"Division by zero attempt");
-        return number/divisor;
+    function divide(uint256 number, uint256 divisor)
+        external
+        pure
+        returns (uint256)
+    {
+        require(divisor > 0, "Division by zero attempt");
+        return number / divisor;
     }
 }
