+++ Issues/8_Bad_Coding_and_Language_Specific/8A_UnmatchedType/UnmatchedType.sol
@@ -4,16 +4,24 @@ pragma solidity 0.8.17;
 //This contract contains an assignment between an uint type and an enum type.
 //Since the range of values for the uint variable is much larger, this could lead to unexpected results.
 contract UnmatchedType {
-    enum Options{ Candidate1, Candidate2, Candidate3 }
+    enum Options {
+        Candidate1,
+        Candidate2,
+        Candidate3
+    }
     mapping(address => Options) private _votes;
-    mapping(Options => uint) private _votesCount;
+    mapping(Options => uint256) private _votesCount;
 
-    function vote(uint option) external{
-        _votes[msg.sender]=Options(option);
+    function vote(uint256 option) external {
+        _votes[msg.sender] = Options(option);
         _votesCount[Options(option)]++;
     }
 
-    function getStatisticsForOption(uint option) external view returns(uint){
+    function getStatisticsForOption(uint256 option)
+        external
+        view
+        returns (uint256)
+    {
         return _votesCount[Options(option)];
     }
 }
