+++ Issues/8_Bad_Coding_and_Language_Specific/8A_UnmatchedType/UnmatchedTypeFixed.sol
@@ -3,18 +3,26 @@ pragma solidity 0.8.17;
 
 //Fix: Check that the value of the uint parameter is in the value range of the enum
 contract UnmatchedTypeFixed {
-    enum Options{ Candidate1, Candidate2, Candidate3 }
+    enum Options {
+        Candidate1,
+        Candidate2,
+        Candidate3
+    }
     mapping(address => Options) private _votes;
-    mapping(Options => uint) private _votesCount;
+    mapping(Options => uint256) private _votesCount;
 
-    function vote(uint option) external{
-        require(option<=uint(Options.Candidate3),"Invalid Option");
-        _votes[msg.sender]=Options(option);
+    function vote(uint256 option) external {
+        require(option <= uint256(Options.Candidate3), "Invalid Option");
+        _votes[msg.sender] = Options(option);
         _votesCount[Options(option)]++;
     }
 
-    function getStatisticsForOption(uint option) external view returns(uint){
-        require(option<=uint(Options.Candidate3),"Invalid Option");
+    function getStatisticsForOption(uint256 option)
+        external
+        view
+        returns (uint256)
+    {
+        require(option <= uint256(Options.Candidate3), "Invalid Option");
         return _votesCount[Options(option)];
     }
 }
