+++ Issues/8_Bad_Coding_and_Language_Specific/8D_DeprecatedFunctions/DeprecatedFunctions.sol
@@ -4,16 +4,18 @@ pragma solidity ^0.6.8;
 //This contract uses "now" to determine the current time
 //Since solidity v 0.7.0, this is deprecated.
 contract DeprecatedFunctions {
-    mapping(address => uint) private _applicants;
-    uint public deadlineTimestamp;
+    mapping(address => uint256) private _applicants;
+    uint256 public deadlineTimestamp;
 
-    constructor(uint _deadlineTimestamp) public
-    {
-        deadlineTimestamp=_deadlineTimestamp;
+    constructor(uint256 _deadlineTimestamp) public {
+        deadlineTimestamp = _deadlineTimestamp;
     }
 
-    function receiveApplication() external{
-        require(now<deadlineTimestamp,"The deadline for applications has been reached");
-        _applicants[msg.sender]=1;
+    function receiveApplication() external {
+        require(
+            now < deadlineTimestamp,
+            "The deadline for applications has been reached"
+        );
+        _applicants[msg.sender] = 1;
     }
 }
