+++ Issues/8_Bad_Coding_and_Language_Specific/8D_DeprecatedFunctions/DeprecatedFunctionsFixed.sol
@@ -3,15 +3,18 @@ pragma solidity 0.8.17;
 
 //Fix: Use an alternative to the deprecated function
 contract DeprecatedFunctionsFixed {
-    mapping(address => uint) private _applicants;
-    uint public deadlineTimestamp;
-    constructor(uint _deadlineTimestamp)
-    {
-        deadlineTimestamp=_deadlineTimestamp;
+    mapping(address => uint256) private _applicants;
+    uint256 public deadlineTimestamp;
+
+    constructor(uint256 _deadlineTimestamp) {
+        deadlineTimestamp = _deadlineTimestamp;
     }
 
-    function receiveApplication() external{
-        require(block.timestamp<deadlineTimestamp,"The deadline for applications has been reached");
-        _applicants[msg.sender]=1;
+    function receiveApplication() external {
+        require(
+            block.timestamp < deadlineTimestamp,
+            "The deadline for applications has been reached"
+        );
+        _applicants[msg.sender] = 1;
     }
 }
