+++ Issues/8_Bad_Coding_and_Language_Specific/8F_UseOfAssembly/UseOfAssembly.sol
@@ -1,42 +1,41 @@
 // SPDX-License-Identifier: MIT
 pragma solidity 0.8.17;
+
 //This contract uses assembly
 contract UseOfAssembly {
-    mapping(address => uint) private _balance;
+    mapping(address => uint256) private _balance;
 
     function deposit() external payable {
-        address sender=msg.sender;
-        uint value=msg.value;
-        assembly{
-
-        mstore(0, sender)
-        mstore(32, _balance.slot)
-        let hash := keccak256(0, 64)
-        sstore(hash, value)        
-        
+        address sender = msg.sender;
+        uint256 value = msg.value;
+        assembly {
+            mstore(0, sender)
+            mstore(32, _balance.slot)
+            let hash := keccak256(0, 64)
+            sstore(hash, value)
         }
     }
 
-     function withdraw() external {
-        address sender=msg.sender;
-        uint toSend;
+    function withdraw() external {
+        address sender = msg.sender;
+        uint256 toSend;
         bool success;
 
-        assembly{
-        mstore(0, sender)
-        mstore(32, _balance.slot)
-        let hash := keccak256(0, 64)
-        toSend := sload(hash) 
+        assembly {
+            mstore(0, sender)
+            mstore(32, _balance.slot)
+            let hash := keccak256(0, 64)
+            toSend := sload(hash)
         }
 
-        require(toSend!=0,"No balance found");
+        require(toSend != 0, "No balance found");
 
-        assembly{
-        mstore(0, sender)
-        mstore(32, _balance.slot)
-        let hash := keccak256(0, 64)
-        sstore(hash,0)
-        success := call(2300,sender,toSend,0,0,0,0x20)
+        assembly {
+            mstore(0, sender)
+            mstore(32, _balance.slot)
+            let hash := keccak256(0, 64)
+            sstore(hash, 0)
+            success := call(2300, sender, toSend, 0, 0, 0, 0x20)
         }
 
         require(success, "Failed to send Ether");
