+++ Issues/8_Bad_Coding_and_Language_Specific/8H_ShadowingVariables/ShadowingParameterVariable.sol
@@ -2,16 +2,23 @@
 pragma solidity 0.8.17;
 
 //The "Split" method declares a local variable with the same name as one fo the parameters.
-contract ShadowingParameterVariable{
-
-    function split(address[] calldata sendTo,uint amountToSend) external payable{
-        require(msg.value>0,"Please also send the currency that is mean to be split");
-        require(sendTo.length>0,"Please specify who the currency will be distributed to");
-        uint amountToSend=msg.value/sendTo.length;
-        for(uint index=0;index<sendTo.length;index++)
-        {
-            bool success=payable(sendTo[index]).send(amountToSend);
-            require(success,"Failed to send currency");
+contract ShadowingParameterVariable {
+    function split(address[] calldata sendTo, uint256 amountToSend)
+        external
+        payable
+    {
+        require(
+            msg.value > 0,
+            "Please also send the currency that is mean to be split"
+        );
+        require(
+            sendTo.length > 0,
+            "Please specify who the currency will be distributed to"
+        );
+        uint256 amountToSend = msg.value / sendTo.length;
+        for (uint256 index = 0; index < sendTo.length; index++) {
+            bool success = payable(sendTo[index]).send(amountToSend);
+            require(success, "Failed to send currency");
         }
     }
 }
