+++ Issues/8_Bad_Coding_and_Language_Specific/8J_DeadCode/DeadCodeFIxed.sol
@@ -3,14 +3,13 @@ pragma solidity 0.8.17;
 
 //Fix: Use the "=" operator as was the intent do perform an assignation.
 contract DeadCodeFIxed {
-    mapping(address => uint) private _participants;
+    mapping(address => uint256) private _participants;
 
-    constructor()
-    {
-        _participants[msg.sender]=0;
+    constructor() {
+        _participants[msg.sender] = 0;
     }
 
-    function deposit(uint option) external{
+    function deposit(uint256 option) external {
         _participants[msg.sender] = option;
     }
 }
